backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
pydantic
alembic
python-dotenv
PyJWT
python-multipart